move proto files to aks-node-controller
