+++ aks-node-controller/app.go
@@ -14,7 +14,7 @@ import (
 	"path/filepath"
 
 	"github.com/Azure/agentbaker/aks-node-controller/parser"
-	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
+	aksnodeconfigv1 "github.com/Azure/agentbaker/aks-node-controller/pkg/gen/aksnodeconfig/v1"
 	"gopkg.in/fsnotify.v1"
 )
