+++ aks-node-controller/helpers/utils_test.go
@@ -7,8 +7,8 @@ import (
 	"reflect"
 	"testing"
 
+	aksnodeconfigv1 "github.com/Azure/agentbaker/aks-node-controller/pkg/gen/aksnodeconfig/v1"
 	"github.com/Azure/agentbaker/pkg/agent/datamodel"
-	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
 )
 
 func Test_getLoadBalancerSKU(t *testing.T) {
