+++ aks-node-controller/parser/helper_test.go
@@ -22,7 +22,7 @@ import (
 	"testing"
 
 	"github.com/Azure/agentbaker/aks-node-controller/helpers"
-	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
+	aksnodeconfigv1 "github.com/Azure/agentbaker/aks-node-controller/pkg/gen/aksnodeconfig/v1"
 	"github.com/Azure/go-autorest/autorest/to"
 )
