+++ aks-node-controller/parser/parser.go
@@ -11,7 +11,7 @@ import (
 	"strings"
 	"text/template"
 
-	aksnodeconfigv1 "github.com/Azure/agentbaker/pkg/gen/aksnodeconfig/v1"
+	aksnodeconfigv1 "github.com/Azure/agentbaker/aks-node-controller/pkg/gen/aksnodeconfig/v1"
 )
 
 var (
