+++ aks-node-controller/proto/aksnodeconfig/v1/api_server_config.proto
