+++ aks-node-controller/proto/aksnodeconfig/v1/auth_config.proto
