+++ aks-node-controller/proto/aksnodeconfig/v1/cluster_config.proto
