+++ aks-node-controller/proto/aksnodeconfig/v1/config.proto
