+++ aks-node-controller/proto/aksnodeconfig/v1/containerd_config.proto
