+++ aks-node-controller/proto/aksnodeconfig/v1/custom_cloud_config.proto
