+++ aks-node-controller/proto/aksnodeconfig/v1/custom_linux_os_config.proto
