+++ aks-node-controller/proto/aksnodeconfig/v1/custom_search_domain_config.proto
