+++ aks-node-controller/proto/aksnodeconfig/v1/gpu_config.proto
