+++ aks-node-controller/proto/aksnodeconfig/v1/http_proxy_config.proto
