+++ aks-node-controller/proto/aksnodeconfig/v1/imds_restriction_config.proto
