+++ aks-node-controller/proto/aksnodeconfig/v1/kube_binary_config.proto
