+++ aks-node-controller/proto/aksnodeconfig/v1/kubelet_config.proto
