+++ aks-node-controller/proto/aksnodeconfig/v1/network_config.proto
