+++ aks-node-controller/proto/aksnodeconfig/v1/teleport_config.proto
