+++ aks-node-controller/proto/aksnodeconfig/v1/tls_bootstrapping_config.proto
